LayoutTests/media/modern-media-controls/media-controller/media-controller-click-on-video-background-to-dismiss-tracks-panel-should-not-toggle-playback-expected.txt

Testing that clicking on a paused video's background to dimiss the tracks panel does not resume it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS media.paused is true

Clicking on tracks button to show tracks panel

Clicking on video background
PASS media.paused is true

PASS successfullyParsed is true

TEST COMPLETE
